refactor: move email service and templates to notification module and update dependencies and configuration
