+++ notification/src/main/resources/templates/email/email-verification.html
