+++ notification/src/main/resources/templates/email/two-factor-authentication.html
