--- shared/src/main/java/pt/estga/shared/services/EmailService.java
@@ -1,9 +1,0 @@
-package pt.estga.shared.services;
-
-import pt.estga.shared.models.Email;
-
-public interface EmailService {
-
-    void sendEmail(Email email);
-
-}
