+++ support/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-pt.estga.support.config.ContactAutoConfiguration
+pt.estga.support.config.SupportAutoConfiguration
